Reorganizes linting utilities under 'collect-data'
Moves several linting-related modules into a 'collect-data' subdirectory
to improve project structure and clarity. Updates import paths across
related files to reflect the new organization.

This change enhances maintainability by grouping data collection
utilities in a more logical location, aligning with project conventions.
